update tests for act flow
And remove 11 redundant tests and combine the files.  The number of tests is getting crazy.
